Use torch.tensor(), set default tensor type based on use_cuda, fix warnings, rename fields and clean up drn_model.py, remove unnecessary logging, and create per-run folder for logs and plots

+++ envs.py
@@ -139,7 +139,7 @@ class AtariEnvironment(Environment):
         self.p = p
 
         self.history_size = history_size
-        self.history = np.zeros([history_size, h, w])
+        self.history = np.zeros([history_size, h, w], dtype='float32')
         self.h = h
         self.w = w
